[Refactor] 사용자의 입력을 받는 기능을 다른 객체로 분리
