+++ BankManagerConsoleApp/BankManagerConsoleApp/Model/Bank.swift
@@ -10,42 +10,16 @@ import Foundation
 struct Bank {
     private let bankManager: BankManager
 
-    init(clerkCount: Int) {
-        self.bankManager = BankManager(clerkCount)
+    init(bankManager: BankManager) {
+        self.bankManager = bankManager
     }
     
-    func open() {
-        while true {
-            self.showMenuScript()
-            
-            guard let input = readLine(),
-                  let userInput = Menu(rawValue: input)
-            else {
-                continue
-            }
-            
-            switch userInput {
-            case .open:
-                self.proceedBanking()
-            case .exit:
-                return
-            case .wrongInput:
-                print(Script.wrongInput)
-            }
-        }
-    }
-    
-    private func proceedBanking() {
+    func proceedBanking() {
         let clientCount = Int.random(in: 10...30)
         
         self.beginReception(count: clientCount)
         self.bankManager.startWork()
         self.showBankSettlement(count: clientCount)
-    }
-    
-    private func showMenuScript() {
-        print(Script.menu)
-        print(Script.inputField, terminator: "")
     }
     
     private func beginReception(count: Int) {
@@ -57,24 +31,5 @@ struct Bank {
     
     private func showBankSettlement(count: Int) {
         print(Script.bankSettlementMessage(count: count, totalWorkTime: self.bankManager.totalWorkTime))
-    }
-}
-
-extension Bank {
-    private enum Menu: String {
-        case open
-        case exit
-        case wrongInput
-        
-        init?(rawValue: String) {
-            switch rawValue {
-            case "1":
-                self = .open
-            case "2":
-                self = .exit
-            default:
-                self = .wrongInput
-            }
-        }
     }
 }
